basic cookiestorage
